Added odometry to rosbag dataset creation process

+++ src/action/scripts/record/data/create_rosbag_dataset.sh
@@ -13,6 +13,7 @@ for bag in $(find ${BAG_DIRECTORY} -name '*.bag.raw'); do
 	INPUT_DATASET_FILE="${bag}"
 	OUTPUT_DATASET_FILE_BAG_PROC_UNBATCHED="${bag}_unbatched.proc"
   OUTPUT_DATASET_FILE_BAG_PROC_MADGWICK="${bag}_madgwick.proc"
+  OUTPUT_DATASET_FILE_BAG_PROC_ODOM="${bag}_odom.proc"
 	OUTPUT_DATASET_FILE_BAG="${bag%.bag.raw}.bag"
 
 	if [ -f $OUTPUT_DATASET_FILE_BAG ]; then
@@ -36,14 +37,20 @@ for bag in $(find ${BAG_DIRECTORY} -name '*.bag.raw'); do
 		exit 1
 	fi
 
-	cp $OUTPUT_DATASET_FILE_BAG_PROC_MADGWICK $OUTPUT_DATASET_FILE_BAG
+  rosrun create create_odometry_rosbag -i $OUTPUT_DATASET_FILE_BAG_PROC_MADGWICK  -o $OUTPUT_DATASET_FILE_BAG_PROC_ODOM
+  if ! [ -f $OUTPUT_DATASET_FILE_BAG_PROC_ODOM ]; then
+		echo "Could not find temporary file ${OUTPUT_DATASET_FILE_BAG_PROC_ODOM}. An error probably occured during conversion."
+		exit 1
+	fi
+
+	cp $OUTPUT_DATASET_FILE_BAG_PROC_ODOM $OUTPUT_DATASET_FILE_BAG
 
 	echo "Compressing rosbag file ${OUTPUT_DATASET_FILE_BAG}"
 	rosbag compress --force --bz2 $OUTPUT_DATASET_FILE_BAG
 	rm -f ${OUTPUT_DATASET_FILE_BAG%.bag}.orig.bag
 
 	echo "Removing all dataset temporary files"
-  rm -f $OUTPUT_DATASET_FILE_BAG_PROC_UNBATCHED $OUTPUT_DATASET_FILE_BAG_PROC_MADGWICK 
+  rm -f $OUTPUT_DATASET_FILE_BAG_PROC_UNBATCHED $OUTPUT_DATASET_FILE_BAG_PROC_MADGWICK $OUTPUT_DATASET_FILE_BAG_PROC_ODOM
 
 done
